Interpolation for Scale and Translate according to CommitStyles
This change updates the Blend operation for scale and translate
transforms, along with CSSFunctionValue calculations for the respective
transforms.

A new Args property is introduced to TransformOperation that keeps track
of how many arguments the primitive scale() and translate() should
have during interpolation.

Follow-up work will include rotate as well as computedStyleMap.

Bug: 1181086
Change-Id: I74c54f47feec66ed5247699543afd1776be8fdcd